requirements.txt
paramiko==3.3.1
netmiko==4.2.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0 